[#30]feature: 훅 추가 및 수정
