+++ src/app/gathering/detail/[id]/_components/FloatingBar.tsx
@@ -1,43 +1,33 @@
 "use client";
 
-import { usePathname } from "next/navigation";
-
 import { Button } from "@/components/ui/Button";
 import { useCancelGathering } from "@/hooks/gathering/useCancelGathering";
 import { useSuspenseGatheringDetail } from "@/hooks/gathering/useGatheringDetail";
 import { useJoinGathering } from "@/hooks/gathering/useJoinGathering";
 import { useIsParticipant } from "@/hooks/gathering/useParticipants";
-import useModalStore from "@/stores/useModalStore";
+import useCopyLink from "@/hooks/useCopyLink";
+import useUserStore from "@/stores/useUserStore";
 
 interface IFloatingBarProps {
   id: string;
 }
 
 const FloatingBar = ({ id }: IFloatingBarProps) => {
-  const pathname = usePathname();
-  const currentUserId = 1667; // 임시 currentUserId: 1667(주최자), 1602(참여자)
-
-  const openAlert = useModalStore(state => state.openAlert);
   const {
     data: { createdBy },
   } = useSuspenseGatheringDetail(id);
+  const user = useUserStore(state => state.user);
+  const currentUserId = user?.id;
+
   const { handleJoinGathering, handleLeaveGathering } = useJoinGathering(
     id,
     currentUserId,
   );
-  const { handleCancelGathering } = useCancelGathering(
-    id,
-    currentUserId === createdBy,
-  );
+  const { handleCancelGathering } = useCancelGathering(id);
+  const { handleCopyLink } = useCopyLink();
 
   const isParticipant = useIsParticipant(id, currentUserId);
   const isOrganizer = currentUserId === createdBy;
-
-  const handleCopyLink = () => {
-    const url = `${window.location.origin}${pathname}`;
-    navigator.clipboard.writeText(url);
-    openAlert("링크가 복사되었습니다.");
-  };
 
   return (
     <div className="fixed bottom-0 left-0 right-0 flex min-h-[84px] items-center border-t-2 border-gray-200 bg-white py-5">
@@ -55,21 +45,21 @@ const FloatingBar = ({ id }: IFloatingBarProps) => {
         {isOrganizer ? (
           <div className="flex gap-2">
             <Button onClick={handleCancelGathering} variant="purple-outline">
-              취소하기
+              {"취소하기"}
             </Button>
             <Button variant="purple" onClick={handleCopyLink}>
-              공유하기
+              {"공유하기"}
             </Button>
           </div>
         ) : (
           <>
             {isParticipant ? (
               <Button onClick={handleLeaveGathering} variant="purple-outline">
-                참여 취소하기
+                {"참여 취소하기"}
               </Button>
             ) : (
               <Button onClick={handleJoinGathering} variant="purple">
-                참여하기
+                {"참여하기"}
               </Button>
             )}
           </>
